Revert "Only tie orders w/users email. Fixes spree/spree#6578"
This code also gets executed if a user is registering or logging in during
checkout: in which case the guest cart will have no email, and won't be
correctly added. It will redirect the user to the address page, not telling him
that his cart was lost and replaced with his last 'signed in cart' -- if the
user isn't careful, they might end up purchasing different items from what they
intended.

This reverts commit 426ebb4b9532125c2c1c3120dfa3a29f37a7bf94.